@@ -1,5 +1,5 @@
 {
-    "name": "laravel/new-cashier",
+    "name": "laravel/cashier",
     "description": "Laravel Cashier provides an expressive, fluent interface to Stripe's subscription billing services.",
     "keywords": ["laravel", "stripe", "billing"],
     "license": "MIT",
@@ -10,18 +10,18 @@
         }
     ],
     "require": {
-        "php": ">=5.5.0",
-        "illuminate/database": "~5.1",
-        "illuminate/support": "~5.1",
+        "php": ">=5.5.9",
+        "illuminate/database": "~5.3",
+        "illuminate/support": "~5.3",
+        "dompdf/dompdf": "^0.6.1",
         "nesbot/carbon": "~1.0",
         "stripe/stripe-php": "~3.0",
-        "symfony/http-kernel": "~2.3"
+        "symfony/http-kernel": "~2.7|~3.0"
     },
     "require-dev": {
-        "dompdf/dompdf": "^0.6.1",
-        "illuminate/routing": "~5.1",
-        "illuminate/http": "~5.1",
-        "illuminate/view": "~5.1",
+        "illuminate/routing": "~5.3",
+        "illuminate/http": "~5.3",
+        "illuminate/view": "~5.3",
         "mockery/mockery": "~0.9",
         "phpunit/phpunit": "~4.0",
         "vlucas/phpdotenv": "~2.0"
@@ -33,7 +33,9 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "7.0-dev"
+            "dev-master": "8.0-dev"
         }
-    }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
