@@ -2,8 +2,24 @@
 
 namespace Laravel\Cashier;
 
+use Exception;
+
 class Cashier
 {
+    /**
+     * The current currency.
+     *
+     * @var string
+     */
+    protected static $currency = 'usd';
+
+    /**
+     * The current currency symbol.
+     *
+     * @var string
+     */
+    protected static $currencySymbol = '$';
+
     /**
      * The custom currency formatter.
      *
@@ -12,15 +28,82 @@ class Cashier
     protected static $formatCurrencyUsing;
 
     /**
+     * Set the currency to be used when billing users.
+     *
+     * @param  string  $currency
+     * @param  string|null  $symbol
+     * @return void
+     */
+    public static function useCurrency($currency, $symbol = null)
+    {
+        static::$currency = $currency;
+
+        static::useCurrencySymbol($symbol ?: static::guessCurrencySymbol($currency));
+    }
+
+    /**
+     * Guess the currency symbol for the given currency.
+     *
+     * @param  string  $currency
+     * @return string
+     */
+    protected static function guessCurrencySymbol($currency)
+    {
+        switch (strtolower($currency)) {
+            case 'usd':
+            case 'aud':
+            case 'cad':
+                return '$';
+            case 'eur':
+                return '€';
+            case 'gbp':
+                return '£';
+            default:
+                throw new Exception('Unable to guess symbol for currency. Please explicitly specify it.');
+        }
+    }
+
+    /**
+     * Get the currency currently in use.
+     *
+     * @return string
+     */
+    public static function usesCurrency()
+    {
+        return static::$currency;
+    }
+
+    /**
+     * Set the currency symbol to be used when formatting currency.
+     *
+     * @param  string  $symbol
+     * @return void
+     */
+    public static function useCurrencySymbol($symbol)
+    {
+        static::$currencySymbol = $symbol;
+    }
+
+    /**
+     * Get the currency symbol currently in use.
+     *
+     * @return string
+     */
+    public static function usesCurrencySymbol()
+    {
+        return static::$currencySymbol;
+    }
+
+    /**
      * Set the custom currency formatter.
      *
      * @param  callable  $callback
      * @return void
      */
-	public static function formatCurrencyUsing(callable $callback)
-	{
+    public static function formatCurrencyUsing(callable $callback)
+    {
         static::$formatCurrencyUsing = $callback;
-	}
+    }
 
     /**
      * Format the given amount into a displayable currency.
@@ -37,9 +120,9 @@ class Cashier
         $amount = number_format($amount / 100, 2);
 
         if (starts_with($amount, '-')) {
-            return '-$'.ltrim($amount, '-');
-        } else {
-            return '$'.$amount;
+            return '-'.static::usesCurrencySymbol().ltrim($amount, '-');
         }
+
+        return static::usesCurrencySymbol().$amount;
     }
 }
